:construction: 线路预计游览时间修改为Double

+++ src/main/java/com/outletcn/app/model/dto/chain/CreateLineRequest.java
@@ -113,7 +113,7 @@ public class CreateLineRequest {
          */
         @NotNull(message = "线路预计游览时间不能为空")
         @ApiModelProperty(value = "线路预计游览时间")
-        private Integer lineExpectTime;
+        private Double lineExpectTime;
 
     }
